Modified the bltmap command

+++ pyxel/tilemap.py
@@ -5,7 +5,7 @@ class Tilemap:
     def __init__(self, width, height):
         self._width = width
         self._height = height
-        self._data = np.zeros((width, height), np.uint16)
+        self._data = np.zeros((width, height), np.uint8)
 
     @property
     def width(self):
